requirements.txt


fastapi==0.105.0
numpy==1.26.4
python-dotenv==1.0.1
annotated-types==0.7.0
anyio==3.7.1
boto3==1.35.14
botocore==1.35.16
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.2.2
filelock==3.16.0
fonttools==4.53.1
fsspec==2024.9.0
idna==3.8
importlib_resources==6.4.5
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psutil==6.0.0
py-cpuinfo==9.0.0
pydantic==2.9.1
pydantic_core==2.23.3
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.2
scipy==1.13.1
seaborn==0.13.2
segment-anything==1.0
six==1.16.0
sniffio==1.3.1
starlette==0.27.0
sympy==1.13.2
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.90
ultralytics-thop==2.0.6
urllib3==1.26.20
zipp==3.20.1
